refactor: Rename pseudo modules overrides
