+++ resources/overrides/index.d.ts
@@ -1,3 +1,3 @@
 import "./jquery.sap.mobile";
 import "./jquery.sap";
-import "./library/index";
+import "./pseudo-modules/index";
