+++ resources/overrides/pseudo-modules/index.d.ts
