+++ resources/overrides/pseudo-modules/sap.collaboration.d.ts
