+++ resources/overrides/pseudo-modules/sap.f.d.ts
