+++ resources/overrides/pseudo-modules/sap.suite.ui.commons.d.ts
