+++ resources/overrides/pseudo-modules/sap.suite.ui.generic.template.d.ts
