+++ resources/overrides/pseudo-modules/sap.suite.ui.microchart.d.ts
