+++ resources/overrides/pseudo-modules/sap.tnt.d.ts
