+++ resources/overrides/pseudo-modules/sap.ui.comp.d.ts
