+++ resources/overrides/pseudo-modules/sap.ui.richtexteditor.d.ts
