+++ resources/overrides/pseudo-modules/sap.ui.table.d.ts
