+++ resources/overrides/pseudo-modules/sap.ui.vbm.d.ts
