+++ resources/overrides/pseudo-modules/sap.ushell.d.ts
