+++ resources/overrides/pseudo-modules/sap.uxap.d.ts
